+++ tests/integration/models/reddit/test_message.py
@@ -88,6 +88,22 @@ class TestMessage(IntegrationTest):
             message = next(self.reddit.inbox.messages())
             message.uncollapse()
 
+    def test_parent(self):
+        self.reddit.read_only = False
+        with self.use_cassette():
+            message = self.reddit.inbox.message("1ay4xyu")
+            parent = message.parent
+            assert isinstance(parent, Message)
+            assert parent.fullname == message.parent_id
+
+    def test_parent__from_inbox_listing(self):
+        self.reddit.read_only = False
+        with self.use_cassette():
+            message = next(self.reddit.inbox.sent(limit=1))
+            parent = message.parent
+            assert isinstance(parent, Message)
+            assert parent.fullname == message.parent_id
+
     @mock.patch("time.sleep", return_value=None)
     def test_reply(self, _):
         self.reddit.read_only = False